Update Material views for land/lake and refector script to import

+++ images/tiler-imposm/start.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 
-# directories to keep the imposm's cache for updating the db
+# Directories to store imposm's cache for updating the DB
 WORKDIR=/mnt/data
 CACHE_DIR=$WORKDIR/cachedir
 DIFF_DIR=$WORKDIR/diff
@@ -11,60 +11,63 @@ PBFFILE="${WORKDIR}/osm.pbf"
 STATEFILE="state.txt"
 LIMITFILE="limitFile.geojson"
 
-# Folder to store the imposm expider files in s3 or gs
+# Folder to store the imposm expirer files in S3 or GCS
 BUCKET_IMPOSM_FOLDER=imposm
-INIT_FILE=/mnt/data/init_done
+INIT_FILE="$WORKDIR/init_done"
 
 PG_CONNECTION="postgresql://$POSTGRES_USER:$POSTGRES_PASSWORD@$POSTGRES_HOST/$POSTGRES_DB"
 
 mkdir -p "$CACHE_DIR" "$DIFF_DIR" "$IMPOSM3_EXPIRE_DIR"
 
-# tracking file
-TRACKING_FILE="$WORKDIR/uploaded_files.log" 
+# Tracking file for uploaded files
+TRACKING_FILE="$WORKDIR/uploaded_files.log"
 [ -f "$TRACKING_FILE" ] || touch "$TRACKING_FILE"
-
 
 # Create config map for imposm
 python build_imposm3_config.py
 
-# Create config file to set variables for imposm
+# Create config file for imposm
+cat <<EOF >"$WORKDIR/config.json"
 {
-    echo "{"
-    echo "\"cachedir\": \"$CACHE_DIR\","
-    echo "\"diffdir\": \"$DIFF_DIR\","
-    echo "\"connection\": \"postgis://$POSTGRES_USER:$POSTGRES_PASSWORD@$POSTGRES_HOST/$POSTGRES_DB\","
-    echo "\"mapping\": \"/osm/config/imposm3.json\","
-    echo "\"replication_url\": \"$REPLICATION_URL\""
-    echo "}"
-} >"$WORKDIR/config.json"
+    "cachedir": "$CACHE_DIR",
+    "diffdir": "$DIFF_DIR",
+    "connection": "postgis://$POSTGRES_USER:$POSTGRES_PASSWORD@$POSTGRES_HOST/$POSTGRES_DB",
+    "mapping": "/osm/config/imposm3.json",
+    "replication_url": "$REPLICATION_URL"
+}
+EOF
 
+function log_message() {
+    echo "$(date +'%Y-%m-%d %H:%M:%S') - $1"
+}
+
+# Function to download the PBF file
 function getData() {
-    ### Get the PBF file from the cloud provider or public URL
-    if [ "$TILER_IMPORT_FROM" == "osm" ]; then
-        if [ ! -f "$PBFFILE" ]; then
-          echo "$PBFFILE does not exist, downloading..."
-          wget "$TILER_IMPORT_PBF_URL" -O "$PBFFILE"
-        fi
+    if [[ "$TILER_IMPORT_FROM" == "osm" && ! -f "$PBFFILE" ]]; then
+        log_message "Downloading PBF file..."
+        wget "$TILER_IMPORT_PBF_URL" -O "$PBFFILE"
     fi
 }
 
+# Function to get the modification date of a file
 getFormattedDate() {
     local file_path="$1"
     if command -v stat >/dev/null 2>&1; then
         local modification_date=$(stat -c %Y "$file_path")
         if [ $? -eq 0 ]; then
             local formatted_date=$(date -d "@$modification_date" "+%Y-%m-%d:%H:%M:%S")
-            echo "Created/Updated date of $file_path: $formatted_date"
+            log_message "Created/Updated date of $file_path: $formatted_date"
         else
-            echo "Error: Unable to get file modification date for file ${file_path}"
+            log_message "Error: Unable to get file modification date for file ${file_path}"
         fi
     else
-        echo "Error: 'stat' command not found. Unable to get file modification date, for file ${file_path}"
+        log_message "Error: 'stat' command not found. Unable to get file modification date, for file ${file_path}"
     fi
 }
 
+# Function to upload expired files
 function uploadExpiredFiles() {
-    echo "Checking for expired files to upload... $(date +%F_%H-%M-%S)"
+    log_message "Checking for expired files to upload..."
 
     # Upload the expired files to the cloud provider
     for file in $(find "$IMPOSM3_EXPIRE_DIR" -type f -cmin -1); do
@@ -73,14 +76,13 @@ function uploadExpiredFiles() {
 
         # Check if the file has already been uploaded
         if grep -Fxq "$file" "$TRACKING_FILE"; then
-            echo "File ${file} has already been uploaded. Skipping..."
+            log_message "File ${file} has already been uploaded. Skipping..."
             continue
         fi
 
         # UPLOAD_EXPIRED_FILES=true to upload the expired files to cloud provider
         if [ "$UPLOAD_EXPIRED_FILES" == "true" ]; then
-            echo "Uploading expired file ${file} to ${AWS_S3_BUCKET}..."
-
+            log_message "Uploading expired file ${file} to ${AWS_S3_BUCKET}..."
             upload_success=false
 
             # AWS
@@ -94,17 +96,18 @@ function uploadExpiredFiles() {
             fi
 
             if [ "$upload_success" = true ]; then
-                echo "$file" >> "$TRACKING_FILE"
-                echo "File ${file} uploaded successfully and recorded."
+                log_message "$file" >> "$TRACKING_FILE"
+                log_message "File ${file} uploaded successfully and recorded."
             else
-                echo "Failed to upload file ${file}. Will retry in the next run."
+                log_message "Failed to upload file ${file}. Will retry in the next run."
             fi
         else
-            echo "Expired files were not uploaded because UPLOAD_EXPIRED_FILES=${UPLOAD_EXPIRED_FILES}"
+            log_message "Expired files were not uploaded because UPLOAD_EXPIRED_FILES=${UPLOAD_EXPIRED_FILES}"
         fi
     done
 }
 
+# Function to upload last state file
 function uploadLastState() {
     # Path to the last.state.txt file
     local state_file="$DIFF_DIR/last.state.txt"
@@ -113,7 +116,7 @@ function uploadLastState() {
 
     # Check if the last.state.txt file exists
     if [ ! -f "$state_file" ]; then
-        echo "No last.state.txt file found at $state_file. Skipping upload."
+        log_message "No last.state.txt file found at $state_file. Skipping upload."
         return
     fi
 
@@ -124,80 +127,95 @@ function uploadLastState() {
     if [ -f "$checksum_file" ]; then
         local previous_checksum=$(cat "$checksum_file")
         if [ "$current_checksum" == "$previous_checksum" ]; then
-            echo "No changes in last.state.txt. Skipping upload."
+            log_message "No changes in last.state.txt. Skipping upload."
             return
         fi
     fi
 
     # Attempt to upload the file to S3
-    echo "Uploading $state_file to S3 at $s3_path..."
+    log_message "Uploading $state_file to S3 at $s3_path..."
     if aws s3 cp "$state_file" "${s3_path}" --acl private; then
         # Update the checksum file after a successful upload
-        echo "$current_checksum" > "$checksum_file"
-        echo "Successfully uploaded $state_file to S3."
+        log_message "$current_checksum" > "$checksum_file"
+        log_message "Successfully uploaded $state_file to S3."
     fi
 }
 
 function updateData() {
+    log_message "Starting database update process..."
 
-    # create views 
-    echo "Create views"
-    psql $PG_CONNECTION -f queries/transport_lines_mviews.sql
-    
+    # Step 1: Ensure required queries are executed
+    log_message "Creating materialized views..."
+    psql "$PG_CONNECTION" -f queries/transport_lines_mviews.sql
+
+    # Step 2: Run any Python-based materialized view updates
+    log_message "Running Python script to update materialized views..."
     python materialized_views.py &
+
+    # Step 3: Start the materialized view refresh in the background
+    log_message "Refreshing materialized views in background..."
+    refreshMaterializedViews &
+
     local local_last_state_path="$DIFF_DIR/last.state.txt"
 
-    ### Update the DB with the new data from minute replication
+    # Step 4: Handle last.state.txt if OVERWRITE_STATE is enabled
     if [ "$OVERWRITE_STATE" = "true" ]; then
-        echo "Overwriting last.state.txt..."
-        echo "timestamp=0001-01-01T00\:00\:00Z 
-        sequenceNumber=${SEQUENCE_NUMBER:-0}
-        replicationUrl=${REPLICATION_URL}" > "$local_last_state_path"
+        log_message "Overwriting last.state.txt..."
+        cat <<EOF > "$local_last_state_path"
+timestamp=0001-01-01T00:00:00Z
+sequenceNumber=${SEQUENCE_NUMBER:-0}
+replicationUrl=${REPLICATION_URL}
+EOF
     fi
 
-    # Run the Imposm update process
+    # Step 5: Run the Imposm update process
+    log_message "Running Imposm update process..."
     if [ -z "$TILER_IMPORT_LIMIT" ]; then
         imposm run \
-        -config "${WORKDIR}/config.json" \
-        -cachedir "${CACHE_DIR}" \
-        -diffdir "${DIFF_DIR}" \
-        -expiretiles-dir "${IMPOSM3_EXPIRE_DIR}" \
-        -quiet &
+            -config "${WORKDIR}/config.json" \
+            -cachedir "${CACHE_DIR}" \
+            -diffdir "${DIFF_DIR}" \
+            -expiretiles-dir "${IMPOSM3_EXPIRE_DIR}" \
+            -quiet &
     else
         imposm run \
-        -config "${WORKDIR}/config.json" \
-        -cachedir "${CACHE_DIR}" \
-        -diffdir "${DIFF_DIR}" \
-        -limitto "${WORKDIR}/${LIMITFILE}" \
-        -expiretiles-dir "${IMPOSM3_EXPIRE_DIR}" \
-        -quiet &
+            -config "${WORKDIR}/config.json" \
+            -cachedir "${CACHE_DIR}" \
+            -diffdir "${DIFF_DIR}" \
+            -limitto "${WORKDIR}/${LIMITFILE}" \
+            -expiretiles-dir "${IMPOSM3_EXPIRE_DIR}" \
+            -quiet &
     fi
 
-    # Continuously upload expired files and last state
+    # Step 6: Continuously upload expired files and last state file to the cloud provider
+    log_message "Starting background upload process..."
     while true; do
-        echo "Upload expired files... $(date +%F_%H-%M-%S)"
+        log_message "Uploading expired files..."
         uploadExpiredFiles
+
+        log_message "Uploading last.state.txt..."
         uploadLastState
+
         sleep 30s
     done
 }
 
 function importData() {
     ### Import the PBF  and Natural Earth files to the DB
-    echo "Execute the missing functions"
+    log_message "Execute the missing functions"
     psql $PG_CONNECTION -f queries/postgis_helpers.sql
 
     if [ "$IMPORT_NATURAL_EARTH" = "true" ]; then
-        echo "Importing Natural Earth..."
+        log_message "Importing Natural Earth..."
         ./scripts/natural_earth.sh
     fi
 
     if [ "$IMPORT_OSM_LAND" = "true" ]; then
-        echo "Import OSM Land..."
+        log_message "Import OSM Land..."
         ./scripts/osm_land.sh
     fi
     
-    echo "Import PBF file..."
+    log_message "Import PBF file..."
     if [ -z "$TILER_IMPORT_LIMIT" ]; then
         imposm import \
             -config $WORKDIR/config.json \
@@ -218,41 +236,66 @@ function importData() {
         -config $WORKDIR/config.json \
         -deployproduction
 
-    # These index will help speed up tegola tile generation
+    log_message "Creating material views and indexes..."
     psql $PG_CONNECTION -f queries/postgis_post_import.sql
+    psql $PG_CONNECTION -f queries/ne_lakes_mviews.sql
     psql $PG_CONNECTION -f queries/land_mviews.sql
     psql $PG_CONNECTION -f queries/transport_lines_mviews.sql
-
-    # To not import again
+    
+    # Create INIT_FILE to prevent re-importing
     touch $INIT_FILE
+}
 
-    # Updata data with minute replication
-    updateData
+function refreshMaterializedViews() {
+    local materialized_views=(
+        "mview_transport_lines_z5_7"
+        "mview_transport_lines_z8_9"
+        "mview_transport_lines_z10_11"
+        "mview_transport_lines_z12_13"
+    )
+
+    while true; do
+        log_message "Refreshing materialized views..."
+
+        for mview in "${materialized_views[@]}"; do
+            log_message "Refreshing $mview..."
+            if psql "$PG_CONNECTION" -c "REFRESH MATERIALIZED VIEW CONCURRENTLY $mview;"; then
+                log_message "Successfully refreshed $mview."
+            else
+                log_message "ERROR refreshing $mview!"
+            fi
+        done
+
+        log_message "Sleeping for 2 minutes before next refresh..."
+        sleep 120
+    done
 }
 
 function countTables() {
     psql $PG_CONNECTION -t -c "SELECT count(*) FROM information_schema.tables WHERE table_schema='public';" | xargs
 }
 
-echo "Connecting to $POSTGRES_HOST DB"
-flag=true
-while $flag; do
-    pg_isready -h $POSTGRES_HOST -p $POSTGRES_PORT >/dev/null 2>&2 || continue
-    # Change flag to false to stop pinging the DB
-    flag=false
-    echo "Run date functions"
-    psql $PG_CONNECTION -f /usr/local/datefunctions/datefunctions.sql
-    echo "Check number of tables in the database"
-    table_count=$(countTables)
-    echo "Check if $INIT_FILE exists"
-    if [[ -f $INIT_FILE || "$table_count" -gt 30 ]]; then
-        updateData
-    else
-        echo "Import PBF data to DB"
-        getData
-        if [ -f $PBFFILE ]; then
-            echo "Start importing the data"
-            importData
-        fi
-    fi
+
+# Wait for PostgreSQL to be ready
+log_message "Waiting for PostgreSQL to be ready..."
+until pg_isready -h "$POSTGRES_HOST" -p "$POSTGRES_PORT" >/dev/null 2>&1; do
+    log_message "PostgreSQL is not ready. Retrying in 2 seconds..."
+    sleep 2
 done
+
+log_message "PostgreSQL is ready! Proceeding with setup..."
+
+# Run date functions
+psql "$PG_CONNECTION" -f /usr/local/datefunctions/datefunctions.sql
+
+# Check the number of tables in the database
+table_count=$(countTables)
+
+# Check if the INIT_FILE exists or if the table count is greater than 30
+if [[ -f "$INIT_FILE" || "$table_count" -gt 30 ]]; then
+    updateData
+else
+    getData
+    importData
+    updateData
+fi
